requirements.txt
multitasking==0.0.11
flask==3.0.0
flask-socketio==5.3.6
openai==1.3.0
yfinance==0.2.31
python-dotenv==1.0.0
httpx==0.27.2
